Have tsconfig.vitest.json reference tsconfig.app.json as it should.
Previously, tsconfig.vitest.json also `include`d app/production code since
- tsconfig.vitest.json extended tsconfig.app.json
- tsconfig.app.json includes production code
- tsconfig.vitest.json didn't specify `include` on its own

The whole point of project references, however, is to *not* do that and keep
production and test code separate.

While `pnpm run type-check` works fine with this change, this seems to produce
an error in VSCode¹ in HelloWorld.spec.ts:

```
File '[…]/tsconfig-test/2024-07-25-vue-setup-with-project-references/src/components/HelloWorld.vue'
is not listed within the file list of project
```

Notably, importing `foo.ts` from the same folder does *not* cause this issue,
so this seems to be Vue-specific.

¹) VSCode with extensions Vue/Volar + Vitest.